[api] Add some initial seeds for games workshop models
This commit adds some initial seed data for games workshop models specifically to boostrap our dataset for manual testing and development.

This seeding functionality is expected to require modifications at a later date to provide an easier to use interface for entering this data and the volume is also expected to increase substantially (this is only a small subset of models in a single faction).