[api] add models#index api endpoint for simple typeahead search
This commit adds a simplistic index api endpoint for typeahead search for models. For now it just takes a single q parameter which represents the query string but this is expected to potentially be built on if requirements for the endpoint appear later.

For now the only use case for this endpoint will be backing a typeahead for model search / selection when creating a recipe.